Fix compute_exec_params
